+++ src/hb/ot_shape_complex_arabic.rs
@@ -3,9 +3,10 @@ use alloc::boxed::Box;
 use super::buffer::{hb_buffer_t, BufferScratchFlags};
 use super::ot_layout::*;
 use super::ot_map::FeatureFlags;
+use super::ot_shape::*;
 use super::ot_shape_complex::*;
 use super::ot_shape_normalize::HB_OT_SHAPE_NORMALIZATION_MODE_AUTO;
-use super::shape_plan::{hb_ot_shape_plan_t, ShapePlanner};
+use super::shape_plan::hb_ot_shape_plan_t;
 use super::unicode::*;
 use super::{hb_font_t, hb_glyph_info_t, hb_mask_t, hb_tag_t, script, Script};
 use crate::hb::feature;
@@ -182,7 +183,7 @@ impl ArabicShapePlan {
     }
 }
 
-fn collect_features(planner: &mut ShapePlanner) {
+fn collect_features(planner: &mut hb_ot_shape_planner_t) {
     // We apply features according to the Arabic spec, with pauses
     // in between most.
     //
