+++ src/hb/ot_shape_complex_hangul.rs
@@ -1,11 +1,12 @@
 use alloc::boxed::Box;
 
+use super::ot_shape::*;
 use crate::hb::buffer::{hb_buffer_t, BufferClusterLevel, BufferFlags};
 use crate::hb::feature;
 use crate::hb::ot_map::*;
 use crate::hb::ot_shape_complex::ComplexShaper;
 use crate::hb::ot_shape_normalize::HB_OT_SHAPE_NORMALIZATION_MODE_NONE;
-use crate::hb::shape_plan::{hb_ot_shape_plan_t, ShapePlanner};
+use crate::hb::shape_plan::hb_ot_shape_plan_t;
 use crate::hb::{hb_font_t, hb_glyph_info_t, hb_mask_t};
 
 pub const HANGUL_SHAPER: ComplexShaper = ComplexShaper {
@@ -65,7 +66,7 @@ impl HangulShapePlan {
     }
 }
 
-fn collect_features(planner: &mut ShapePlanner) {
+fn collect_features(planner: &mut hb_ot_shape_planner_t) {
     planner
         .ot_map
         .add_feature(feature::LEADING_JAMO_FORMS, FeatureFlags::empty(), 1);
@@ -77,7 +78,7 @@ fn collect_features(planner: &mut ShapePlanner) {
         .add_feature(feature::TRAILING_JAMO_FORMS, FeatureFlags::empty(), 1);
 }
 
-fn override_features(planner: &mut ShapePlanner) {
+fn override_features(planner: &mut hb_ot_shape_planner_t) {
     // Uniscribe does not apply 'calt' for Hangul, and certain fonts
     // (Noto Sans CJK, Source Sans Han, etc) apply all of jamo lookups
     // in calt, which is not desirable.
