+++ src/hb/ot_shape_complex_khmer.rs
@@ -1,5 +1,6 @@
 use alloc::boxed::Box;
 
+use super::ot_shape::*;
 use super::ot_shape_normalize::hb_ot_shape_normalize_context_t;
 use crate::hb::buffer::hb_buffer_t;
 use crate::hb::feature;
@@ -7,7 +8,7 @@ use crate::hb::ot_map::FeatureFlags;
 use crate::hb::ot_shape_complex::*;
 use crate::hb::ot_shape_complex_indic::{category, position};
 use crate::hb::ot_shape_normalize::HB_OT_SHAPE_NORMALIZATION_MODE_COMPOSED_DIACRITICS_NO_SHORT_CIRCUIT;
-use crate::hb::shape_plan::{hb_ot_shape_plan_t, ShapePlanner};
+use crate::hb::shape_plan::hb_ot_shape_plan_t;
 use crate::hb::unicode::{CharExt, GeneralCategoryExt};
 use crate::hb::{hb_font_t, hb_glyph_info_t, hb_mask_t, hb_tag_t};
 
@@ -121,7 +122,7 @@ impl KhmerShapePlan {
     }
 }
 
-fn collect_features(planner: &mut ShapePlanner) {
+fn collect_features(planner: &mut hb_ot_shape_planner_t) {
     // Do this before any lookups have been applied.
     planner.ot_map.add_gsub_pause(Some(setup_syllables));
     planner.ot_map.add_gsub_pause(Some(reorder));
@@ -291,7 +292,7 @@ fn reorder_consonant_syllable(
     }
 }
 
-fn override_features(planner: &mut ShapePlanner) {
+fn override_features(planner: &mut hb_ot_shape_planner_t) {
     // Khmer spec has 'clig' as part of required shaping features:
     // "Apply feature 'clig' to form ligatures that are desired for
     // typographical correctness.", hence in overrides...
