+++ src/hb/ot_shape_complex_myanmar.rs
@@ -1,3 +1,4 @@
+use super::ot_shape::*;
 use crate::hb::buffer::hb_buffer_t;
 use crate::hb::feature;
 use crate::hb::ot_map::FeatureFlags;
@@ -7,7 +8,7 @@ use crate::hb::ot_shape_normalize::{
     HB_OT_SHAPE_NORMALIZATION_MODE_COMPOSED_DIACRITICS_NO_SHORT_CIRCUIT,
     HB_OT_SHAPE_NORMALIZATION_MODE_NONE,
 };
-use crate::hb::shape_plan::{hb_ot_shape_plan_t, ShapePlanner};
+use crate::hb::shape_plan::hb_ot_shape_plan_t;
 use crate::hb::{hb_font_t, hb_glyph_info_t, hb_tag_t};
 
 pub const MYANMAR_SHAPER: ComplexShaper = ComplexShaper {
@@ -140,7 +141,7 @@ impl hb_glyph_info_t {
     }
 }
 
-fn collect_features(planner: &mut ShapePlanner) {
+fn collect_features(planner: &mut hb_ot_shape_planner_t) {
     // Do this before any lookups have been applied.
     planner.ot_map.add_gsub_pause(Some(setup_syllables));
